add,refactor: create_timer_app_page
