+++ mobile/lib/component/local/todo_and_timer_app/todo/todo_card/todo_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../shared/token/color/color.dart';
+import '../../../../shared/token/color/color.dart';
 
 class TodoCard extends StatelessWidget {
   const TodoCard({super.key, required this.text});
